+++ cassandra/query.py
@@ -8,10 +8,10 @@ from datetime import datetime, timedelta
 import struct
 import time
 
-from cassandra import ConsistencyLevel
+from cassandra import ConsistencyLevel, OperationTimedOut
 from cassandra.cqltypes import unix_time_from_uuid1
 from cassandra.decoder import (cql_encoders, cql_encode_object,
-                               cql_encode_sequence)
+                               cql_encode_sequence, named_tuple_factory)
 
 import logging
 log = logging.getLogger(__name__)
@@ -45,10 +45,8 @@ class Statement(object):
 
     _routing_key = None
 
-    def __init__(self, retry_policy=None, tracing_enabled=False,
-                 consistency_level=None, routing_key=None):
+    def __init__(self, retry_policy=None, consistency_level=None, routing_key=None):
         self.retry_policy = retry_policy
-        self.tracing_enabled = tracing_enabled
         if consistency_level is not None:
             self.consistency_level = consistency_level
         self._routing_key = routing_key
@@ -264,10 +262,9 @@ class BoundStatement(Statement):
                     expected_type = col_type
                     actual_type = type(value)
 
-                    err = InvalidParameterTypeError(col_name=col_name,
-                                                    expected_type=expected_type,
-                                                    actual_type=actual_type)
-                    raise err
+                    message = ('Received an argument of invalid type for column "%s". '
+                               'Expected: %s, Got: %s' % (col_name, expected_type, actual_type))
+                    raise TypeError(message)
 
         return self
 
@@ -345,24 +342,6 @@ class TraceUnavailable(Exception):
     pass
 
 
-class InvalidParameterTypeError(TypeError):
-    """
-    Raised when a used tries to bind a prepared statement with an argument of an
-    invalid type.
-    """
-
-    def __init__(self, col_name, expected_type, actual_type):
-        self.col_name = col_name
-        self.expected_type = expected_type
-        self.actual_type = actual_type
-
-        values = (self.col_name, self.expected_type, self.actual_type)
-        message = ('Received an argument of invalid type for column "%s". '
-                   'Expected: %s, Got: %s' % values)
-
-        super(InvalidParameterTypeError, self).__init__(message)
-
-
 class QueryTrace(object):
     """
     A trace of the duration and events that occurred when executing
@@ -433,9 +412,13 @@ class QueryTrace(object):
         attempt = 0
         start = time.time()
         while True:
-            if max_wait is not None and time.time() - start >= max_wait:
+            time_spent = time.time() - start
+            if max_wait is not None and time_spent >= max_wait:
                 raise TraceUnavailable("Trace information was not available within %f seconds" % (max_wait,))
-            session_results = self._session.execute(self._SELECT_SESSIONS_FORMAT, (self.trace_id,))
+
+            session_results = self._execute(
+                self._SELECT_SESSIONS_FORMAT, (self.trace_id,), time_spent, max_wait)
+
             if not session_results or session_results[0].duration is None:
                 time.sleep(self._BASE_RETRY_SLEEP * (2 ** attempt))
                 attempt += 1
@@ -448,10 +431,24 @@ class QueryTrace(object):
             self.coordinator = session_row.coordinator
             self.parameters = session_row.parameters
 
-            event_results = self._session.execute(self._SELECT_EVENTS_FORMAT, (self.trace_id,))
+            time_spent = time.time() - start
+            event_results = self._execute(
+                self._SELECT_EVENTS_FORMAT, (self.trace_id,), time_spent, max_wait)
             self.events = tuple(TraceEvent(r.activity, r.event_id, r.source, r.source_elapsed, r.thread)
                                 for r in event_results)
             break
+
+    def _execute(self, query, parameters, time_spent, max_wait):
+        # in case the user switched the row factory, set it to namedtuple for this query
+        future = self._session._create_response_future(query, parameters, trace=False)
+        future.row_factory = named_tuple_factory
+        future.send_request()
+
+        timeout = (max_wait - time_spent) if max_wait is not None else None
+        try:
+            return future.result(timeout=timeout)
+        except OperationTimedOut:
+            raise TraceUnavailable("Trace information was not available within %f seconds" % (max_wait,))
 
     def __str__(self):
         return "%s [%s] coordinator: %s, started at: %s, duration: %s, parameters: %s" \
@@ -495,7 +492,10 @@ class TraceEvent(object):
         self.description = description
         self.datetime = datetime.utcfromtimestamp(unix_time_from_uuid1(timeuuid))
         self.source = source
-        self.source_elapsed = timedelta(microseconds=source_elapsed)
+        if source_elapsed is not None:
+            self.source_elapsed = timedelta(microseconds=source_elapsed)
+        else:
+            self.source_elapsed = None
         self.thread_name = thread_name
 
     def __str__(self):